Tag image with '*-main' tags after running tests
This would allow us to use Renovate to update the references to stable
images. The image tags can be ordered. The new tag format is:

    {commit-timestamp}{short-commit-sha}-main

The only problem would be if two commits have the same committed
date/time (the resolution is one second). In that case, it could happen
that the older commit tag gets applied later, because the push pipelines
can run in parallel.

Also, tags images built from Pull Requests with `pr{PR_ID}`.

JIRA: RHELWF-12927