Drop "version" attribute from rules
The "version" is currently not used and it is unclear if it would be
useful in the future. Also, it can be confusing, because it seems to
indicate version of the rule itself, but, in fact, it is the schema
version for the rules.

Dropping the value simplifies the definitions a bit.