Support kerberos Product Pages authentication for local runs
This enables users to run ReTaSC locally to test custom rules, and
simplifies development testing.

It requires separate command (curl) to authenticate with Product Pages
and storing the cookies file. The cookie file can be then used by
ReTaSC.

    kinit
    export RETASC_PRODUCT_PAGES_COOKIES=$PWD/pp_cookies.txt
    curl --negotiate --location-trusted -d "" -sLc $RETASC_PRODUCT_PAGES_COOKIES \
        https://product_pages.example.com/oidc/authenticate

An alternative would be to use requests-gssapi library to handle the
authentication, but it requires installing it in the container image
with krb5 development libraries and also running kinit or passing the
kerberos ticket to the container.

JIRA: RHELWF-13514